% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics_styling.R
\docType{data}
\name{dezim_colors}
\alias{dezim_colors}
\title{Vector of colors for DeZIM Plots}
\format{
\subsection{\code{dezim_colors}}{

A vector of colors to be used with fill/color functions for DeZIM graphs
\describe{
\item{color}{Colors found in DeZIM style guide}
...
}
}
}
\usage{
dezim_colors
}
\description{
Vector of colors for DeZIM Plots
}
\keyword{datasets}
