% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics_styling.R
\name{dezim_style}
\alias{dezim_style}
\title{DeZIM Graphics Style}
\usage{
dezim_style(
  title_size = 10.5,
  sub_title_size = 9,
  x_axis_title_size = 10,
  y_axis_title_size = 10,
  x_axis_text_size = 9,
  y_axis_text_size = 9,
  legend_text_size = 9
)
}
\arguments{
\item{title_size}{Size of title}

\item{sub_title_size}{Size of subtitle}

\item{x_axis_title_size}{size of x axis title}

\item{y_axis_title_size}{size of y axis title}

\item{x_axis_text_size}{size of x axis text}

\item{y_axis_text_size}{size of y axis text}

\item{legend_text_size}{Size of text in legend}
}
\value{
No return, a theme to add to ggplot2 graphs
}
\description{
DeZIM Graphics Style
}
\examples{
library(surveyexplorer)

#add "+ dezim_style to ggplot2 graphs
surveyexplorer::matrix_table(berlinbears, dplyr::starts_with('p_')) +
dezim_style()

#use dezim_colors for functions that take colors
 surveyexplorer::matrix_freq(berlinbears, dplyr::starts_with('p_')) +
 dezim_style() + ggplot2::scale_fill_manual(values = dezim::dezim_colors)



}
