% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics_styling.R
\name{dezim_style}
\alias{dezim_style}
\title{DeZIM Graphics Style}
\usage{
dezim_style(
  title_size = 10.5,
  sub_title_size = 9,
  x_axis_title_size = 10,
  y_axis_title_size = 10,
  x_axis_text_size = 9,
  y_axis_text_size = 9,
  legend_text_size = 9
)
}
\arguments{
\item{title_size}{Size of title}

\item{sub_title_size}{Size of subtitle}

\item{x_axis_title_size}{size of x axis title}

\item{y_axis_title_size}{size of y axis title}

\item{x_axis_text_size}{size of x axis text}

\item{y_axis_text_size}{size of y axis text}

\item{legend_text_size}{Size of text in legend}
}
\value{
No return, a theme to add to ggplot2 graphs
}
\description{
DeZIM Graphics Style
}
\examples{
library(ggplot2)
mtcars$gear <- as.factor(mtcars$gear)

p1 <- ggplot(mtcars) +
geom_point(aes(x = wt, y = mpg, colour = gear)) +
labs(
title = "Fuel economy declines as weight increases",
 subtitle = "(1973-74)",
  x = "Weight (1000 lbs)",
  y = "Fuel economy (mpg)",
  colour = "Gears"
  )

#Default graph
 p1

 \dontrun{
 #Examples are not run since test machine does not have Calibri installed
#use `+ dezim_style()` to apply theme
 p1 +
 dezim_style()

#use `dezim_colors` for fill/colors
 p1 +
 dezim_style() +
 scale_color_manual(values = dezim_colors)
}



}
